Fix lint warnings

+++ axmol/rhi/d3d/RenderPipelineD3D.h
@@ -17,6 +17,6 @@ public:
     ID3D11Device* _device         = nullptr;
     ID3D11DeviceContext* _context = nullptr;
 
-    tsl::robin_map<uint64_t, ComPtr<ID3D11BlendState>> _blendCache;
+    tsl::robin_map<uint32_t, ComPtr<ID3D11BlendState>> _blendCache;
 };
 }  // namespace ax::rhi::d3d
